mods

+++ baselines/common/tf_util.py
@@ -331,10 +331,10 @@ def dense(x, size, name, weight_init=None, bias=True):
         return ret
 
 def dense_wparams(x, size, name, weight_init=None, bias=True):
-    w = tf.get_variable(name + "/w", [x.get_shape()[1], size], initializer=weight_init)
+    w = tf.get_variable(name + "/w", [x.get_shape()[1], size], initializer=weight_init, dtype=tf.float64)
     ret = tf.matmul(x, w)
     if bias:
-        b = tf.get_variable(name + "/b", [size], initializer=tf.zeros_initializer())
+        b = tf.get_variable(name + "/b", [size], initializer=tf.zeros_initializer(), dtype=tf.float64)
         return ret + b, w,b
     else:
         return ret,w
